requirements.txt
alembic==1.7.3
backcall==0.2.0
bcrypt==3.2.0
click==8.0.1
decorator==5.0.9
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gunicorn==20.1.0
idna==3.2
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
Pillow==8.3.2
prompt-toolkit==3.0.17
psycopg2==2.9.1
pycparser==2.20
Pygments==2.8.1
SQLAlchemy==1.4.23
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==2.0.1
WTForms==2.3.3